[netstack] Refactor test_watcher
Refactor `test_watcher` to make use of new facilities that've been
adopted since the test was written to make it cleaner.

Relax the condition that address changes must occur after the online
property changes in preparation for an upcoming change in which this
will no longer be guaranteed.

Test: fx test --test-filter test_watcher netstack-interfaces-fidl-test
Multiply: netstack-interfaces-fidl-test
Change-Id: Icbf53a14cf507dba846206022d2f670deeddc584
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700470
Reviewed-by: Nate Hurley <nahurley@google.com>
Commit-Queue: Tony Gong <gongt@google.com>