[ffx] Add trait to engines to access EmuConfig
One asymmetry with the emulator engines is that they require an
EmulatorConfiguration as input, but have no way to access that
configuration once it's provided. This hasn't been a problem before now,
but as we prepare to enable editing of the emulator's configuration
pre-launch, we need to either return the configuration from the stage
command or provide a mechanism for mutably referencing the config within
the engine.

This CL moves two funtions to the EmulatorEngine trait, emu_config and
emu_config_mut, which return immutable or mutable references to the
object, respectively. These functions were previously part of the
QemuBasedEngine trait, but moving them this way makes them more widely
accessible - in particular from the start subcommand entry point, where
we will need them for the editor functionality to come.

Bug: 98268
Change-Id: I7857cb88baec2e153eb54df4f101c862733f027d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700944
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>