[dart][test] Immediately report when the real runner closes controller
This updates the dart wrapper runner to immedately close the client
half of the controller channel when running a non-test component and
the server channel is closed.
Channel closure in the ComponentController is semantically meaningful -
it indicates that the component has completed running.

In the original change, we didn't forward channel closure as we found
that fuchsia.test.Suite continues to run for a Dart test after the
channel is closed. Waiting for a Stop or Kill signal gave us a
signal which indicated when test_manager had finished draining
fuchsia.test.Suite. This had the side effect that calling Stop on a
non-test dart component would time out since we didn't forward channel
closure. To correct this case, we check that the component is a test by
looking for presence of an is_test=true argument in component start
info, and propogating the stop signal if the component is not a test.

Testing: fx test --count 1000 dart_realm_builder_unittests
fx test inspect-validator-test-dart against patchset 13 of
fxrev.dev/698139, which calls stop on a non-test Dart component.

Multiply: dart_realm_builder_unittests

Change-Id: I3c084a634390f2f87b66823be159fcbd2434cb75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701143
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Naud Ghebre <naudzghebre@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>