[ffx] Change timeout message to something nicer
Since the emulator landed in March '22, the most common statement in
chats and bug reports has been akin to  "My emulator timed out". This
statement comes from the message that's printed whenever the emulator
fails to respond to a network request from the Daemon, which can be a
symptom of a problem, but can also mean the process is taking longer
than expected. It doesn't tell the user anything useful, and it doesn't
provide any information useful for debugging.

In this CL, we replace the "timed out" message with something more
useful for both the user and the devs helping them troubleshoot. The
wording around "timing out" is removed, and a more explicit message
about network requests is put in its place. Following that, we list
several of the pieces of information that we commonly ask users when
debugging these situations, including the emulator's networking mode,
acceleration mode, and pid, as well as common next steps the user can
execute on their own before reaching out for help.

Change-Id: I0d4b96ab8f79be8edd1142f4dbdb7cad0e8dfdd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701230
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>