[fidlgen_banjo][rust] Fix new compiler warning
Some old code was formatting a named variable but not reference the
name in the formatting string. This change adds the identifier to the
formatting string, but does not affect the behavior of the formatting. The new compiler warning was introduced in https://github.com/rust-lang/rust/issues/98466

Bug: 104577

Change-Id: I7a19d1aa23a12fd116dfb5488f0c233cf24e3cb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701186
Commit-Queue: Paul Faria <paulfaria@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>