[fdf][compat] Improve devfs export log messages.
Previously we would log "Created <path>" prior to actually exporting the
device, and if exporting failed we would log an error that didn't
mention which device failed to be exported.

This CL changes it so that we log "Created <path>" *after* export
succeeds, and "Failed to export <path>" if export fails.

Change-Id: Ie1fd6d370708edb27c4fdf512f66eb5d3505c4ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700963
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>