[security][sc] gn templates for structured config goldens
This change structures the current strategy for verifying
structured configuration against goldens by providing gn templates
for the "verify against golden based on build type" pattern:

- verify_file_by_build_type(): Verify an arbitrary file against
  goldens that may differ by non-eng/eng build type;
- verify_structured_config_by_build_type(): Verify a structured
  configuration against goldens that may differ by non-eng/eng
  build type.

The one use of this pattern is updated to use one of the templates.

In addition, the `--eng-golden` parameter for the underlying script
is made optional because verifying the eng build type is usually
not security-relevant. Finally, a README is introduced to describe
the function of files in `//src/security/policy/build/`, including
those used for the above-mentioned pattern.

Bug: 104220
Change-Id: I8f56d66dae63003ea2d19d088b927b91b776a9cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700078
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>