[core] Only offer DRM protocols when supported.
Change session_manager to only require DRM protocols when the session
itself requires them. Then use `source_availability: "unknown"` to
only offer them when the `#playready-cdm` component is present as a
child of `core.cm`.

The end result is that we get an error if these protocols were
required by the session (or a child thereof), but not provided by
`core`.

Bug: 103977

Change-Id: I484f246e91d175156e59938ba38e0c5d03fe8b0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700086
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Wez <wez@google.com>