[ftl] Add diagnosis for running out of bad blocks.
Reports a failure symptom for when the number of marked bad blocks on an
ftl volume is at or exceeds (though it should never be able to exceed)
the number of spare blocks. Once in this state another bad block cannot
be marked bad and the ftl will fail out when attempting to do so. It is
the likely cause of failure when run on the device for an image that
otherwise appears intact.

Tested: On a known failing image.

Change-Id: Ia7f5c43bd555804de6db1b9002ce520ab0358af1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701557
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>