[network] Terminate Rx goroutines on socket close
An edge case arises on the Fast UDP Rx path when the zircon socket is
almost full and the last datagram payload can't fit. When the socket is
closed under these circumstances, the Rx goroutine needs to exit;
otherwise, it will hang indefinitely.

In order to test this case:

- Share testing utils between {stream,dgram}socket_test.cc.
- Support getting Tx and Rx capacities for file descriptors backed by
  `fposix_socket::DatagramSocket`s.
- Log `ErrConnectionRefused` errors with Debug severity. These errors
  are expected when the gVisor endpoint's local peer is closed and we
  shouldn't fail tests when we see them.

Fixed: 104490
Bug: 104491
Change-Id: I785868b34cbd370dd031cf03d49b270b0abafa66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700806
Reviewed-by: Joon-Hyuk Han <hanjh@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>