[perfmon] Propagate zx_status
We drop the zx_status_t we are given from the perfmon controller and
return it as a bool. This means that we don't have a way to determine
_why_ perfmon failed to start if it did except by reading the logs. This
changes the status results to now be a zx::status so that we keep and
error and are force to check/propagate it which should be a bit safer.

This also adds an integration test to ensure we can start and stop
perfmon if available.

Bug: 101746
Change-Id: Iecef6ea78d3c8fb8aedbe886bcaf92148c6d28b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698404
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>