[acpi] Make device DAG reflect ACPI tables.
In I8fb14cae5e7bbbb05bb5ef58d8272336558c05fb we made all devices
children of the root ACPI device to work around the fact that we cannot
add children to a node that a driver will bind to.

It is possible though for multiple devices with the same name to exist
at different points in the ACPI namespace, and recently we've started
noticing this. This CL reverts to the state prior to
I8fb14cae5e7bbbb05bb5ef58d8272336558c05fb but continues to respect the
driver framework's constraints by adding a new `-passthrough` child
node to each ACPI device, which is what drivers bind to.

Test: boot QEMU, Atlas and check hardware works.
Test: check Atlas on DFv2.
Change-Id: I81701f2a50a9b122b0c6e377ccdb2558d998401f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694727
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>