[virtualization][vsock] Deflake test
This test involves sending a large packet to the device and then reading
a byte from the socket. As the device is out of process, we need to poll
the socket before continuing the test to know that the packet has been
processed.

Multiply: fuchsia-pkg://fuchsia.com/device_tests#meta/device_tests.cm
Bug: 104599,104553
Change-Id: I9afb17e3f229cbdea7a192795a583a6ade4f2d04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701390
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>