[ffx] Cleanup compatibility code for architecture
CL 674057 changed the host.architecture field to use an enumeration,
which modified its serialized form. To account for this, we introduced
compatibility code to the tool to handle both serialized forms, to make
sure we didn't break running emulators that still had the old serialized
data.

It's been about 3 months since that CL landed, so this CL removes the
compatibility code, such that we now only accept the new form based on
the enumeration type. It's expected that any emulators that were running
when that CL landed have been torn down, and the SDKs thah held the old
version of the tool have been upgraded by now, so no more instances of
the old serialized data should still be around.

Bug: 99291
Change-Id: I40f3fd79d8c6105ebca15bcabf6b96fa6caf28df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701543
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>