Reland "[audio][drivers][hda] Changes to deprecate intel_hda::Status/StatusOr and use zx::status instead"
This is a reland of commit 4d50a61614359237dc021bac87b59f39ba272a65

The fix for fxrev.dev/696902 exposed (in the asan build) an address access issue caused by race condition between the Audio driver and the DestructWithQueuedThread test thread of intel-hda-controller-test.
The DestructWithQueuedThread and Audio driver use an object of type DspChannel. The test races to destructs this common object while the driver is still using it.
One way to handle this issue, is to call Shutdown() of the object to gracefully signal the driver, wait for the driver thread to join and finally call the reset() to destruct the object.
But this change in code will make this test redundant as  ShutdownWithQueuedSend tests the same sequence.
This new CL removes the DestructWithQueuedThread test as its no longer
needed.

Fixed: 86895
Test: fx test intel-dsp-stream-test intel-hda-stream-test intel-hda-controller-fuzzers
      intel-hda-controller-test intel-hda-test intel-hda-utils-test-test
Test: signal_generator on Atlas
Test: intel-hda-controller-test under asan repeated 9999 times

Original change's description:
>
>Fixed: 86895
>Test: fx test intel-dsp-stream-test intel-hda-stream-test intel-hda-controller-fuzzers
>      intel-hda-controller-test intel-hda-test intel-hda-utils-test-test
>
>      Test with signal_generator on Atlas
>
>Change-Id: I6626be030238901efb5d4485011be84408a29072
>Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696902
>Commit-Queue: Harsha Priya N V <harshanv@google.com>
>Reviewed-by: Andres Oportus <andresoportus@google.com>

Change-Id: I7cf72be2ca48d83a937a8c6a6f097a376ea35d74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701687
Commit-Queue: Harsha Priya N V <harshanv@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>