[ot-radio] Update pending tid log level
The ot-radio driver is tracking the pending TIDs while resetting, and
generate an error message if outstanding request haven't receive the
response when resetting the ot-radio driver.

This is helpful to triage issues related to reset in Thread 1.1
releases, since the force reset of a radio means the thread stack is
crashed and needs to brought up again.

However, when implementing the Thread 1.3 Backbone IF feature,
lowpan-ot-driver may be forced to restart when the Backbone IF is found,
which is a valid case and should not generate an error message. Hence I
am changing the log level from error to warning.

Bug: 104100
Change-Id: Iaaac50e1ffd7a331c9ef02c2f34743d84c8272a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701628
Reviewed-by: Robert Quattlebaum <rquattle@google.com>
Commit-Queue: Jiaming (Charlie) Wang <jiamingw@google.com>