[cmxtinction] Convert trace providers to CFv2
CFv1 is deprecated, time to move to CFv2!

This moves cpuperf_provider and ktrace_provider to CFv2. Instead of
having providers in a config file, we instead configure them as eager
children of the trace manager and route the the capabilities they need.

This also moves a few integration tests to use the gtest runner. It
makes some changes to when the tests bail out if there is not perfmon.
If we return EXIT_SUCCESS but haven't run any tests, we crash the test
runner. So instead, we start the test and immediately return.

Fixed: 103220
Change-Id: I0a6501ebf97b07cf13528b63c869d45027f14527
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693372
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>