[libc] Do not build lsan-test with the hwasan variant toolchain
Otherwise we will run into many undefined symbol link errors when
building this test because the hwasan runtime doesn't come shipped
with lsan. It could be possible in the future for this to happen,
but that shouldn't be a blocker now, so what we can do is just ensure
this isn't built with hwasan.

Bug: 57112
Change-Id: Ia915192dc7067485aca3b75de3130fe2e628ab4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700131
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>