[ffx][pkg] Tests should specify abi revision
When the ABI revision changes, the package meta.far merkle (and
theoretically the meta.far size) can change. To avoid these tests breaking
when we update the ABI revision, this changes these tests to pin to a
specific ABI revision in most places. In addition, this adds a test to
check that we can still target the latest ABI, but we automatically compute
the meta.far merkle and size, rather than hardcoding the values.

Change-Id: I1af439ba9dfae013ed076f32f68b0eeb71ad87c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702142
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>