[roll] Update platform version from 8 to 9
API golden files generated with:
$ ... update build/config/fuchsia/platform_version.json ...
$ mkdir sdk/history/9
$ fx set core.x64 --with /* merged sdk */ --args=api_compatibility_testing=false
$ fx build
$ find out/default -type f -wholename "*/9/fuchsia.*.api_summary.json" -exec cp {} sdk/history/9/ \;

The change to src/lib/assembly/update_package/src/update_package.rs is
required because the hashes are currently based on the latest ABI
revision and this CL changes it. The assembly team will work on a long
term fix to address this.

Change-Id: I4eed5a6ac8c73ebca975224c2c2b99161e7b912e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701141
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Sébastien Marchand <sebmarchand@google.com>