[toolchain] Exclude hwasan variant from being mixed with linux toolchains
We hit a lot of build errors when attempting to build core.arm64-hwasan
because many linux binaries built for starnix would have a hwasan
variant of them built. This isn't surprising because we currently don't
ship the hwasan runtime for linux.

This fixes the issue by introducing a `fuchsia-only` tag that ensures a
variant will only be applied to toolchains for building fuchsia targets
and adding that tag to the hwasan variant.

Bug: 57112
Change-Id: I0d514f9c16f7a0ff33eb9808c50dfe5a0142adf7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700075
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>