[virtualization][vsock] Update device state change synchronization
When handling TX the device acquires a write lock on the state as the
state may change, and to do that aborts any futures holding read locks.
These aborts themselves need an async lock so that they're not
re-created while a writer is waiting.

Tested:
- Booted termina with this device
- fx test virtualization_vsock_tests (debian + termina)

Bug: 97355
Change-Id: I0d920ac87a91d605ada954ab857a105dd9dd8eec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701385
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>