[recovery] Shrink /data partition size in recovery ramdisk on chromebook-x64 board
The ramdisk that recovery uses was too small to fit the 4GiB configured for the /data partition on the chromebook-x64 board.

This changes the arguments used to set the partition sizes in chromebook-x64.gni from build arguments to fshost config arguments so that they don't affect the recovery fshost config.

Change-Id: Ifdba64b03d9031c079dda04525dbbda925ea7418
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698240
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Adam Chyb <achyb@google.com>
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>