[ffx] Generate ssh keys if none found
This checks the `ssh.pub` and `ssh.priv` configuration values, and if
no existing keys are found, a new ed25519 key is generated and saved at
the location of the first value configured. This is typically
$HOME/.ssh/fuchsia_ed25519 and fuchsia_authorized_keys.

Typical usage would be to check for the existence when initializing
a device via flashing/paving, or starting an emulator instance. This
is preferred since generating the private key when attempting to access
a device already initialized is pointless.

Also addressed issues from `fx clippy`.

Fixed: 97603, 76032
Change-Id: Ie5a839b9ea3104808e9fc3bf57d5285f3f28d539
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699186
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>