[ffx] Refactor engine code to reflect dependencies
The Fuchsia emulator launcher supports two engines, "qemu" and "femu",
which are both based on different versions of the Qemu emulation system.
As such, they share a significant amount of code, but they have a few
differences which require separate implementations. Previously, the
majority of the shared code lived in the engines/qemu/ directory, along
with the leaf of the "qemu" implementation, while the "femu" leaf was
in engines/femu and had a dependency on its peer directory.

This CL shifts the files within the directory structure to better
reflect how the files work together. The root is now engines/qemu_based,
and it contains the shared code as mod.rs as well as subdirectories for
qemu and femu which contain their respective leaf implementations. This
has no functional impact on the program, but should make it easier to
find the implementations for folks new to the project.

Change-Id: If8800bab992c671951706a09f76bef9e2ba97149
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702030
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>