[mixer_service] Refactor GainControl
This change refactors the GainControl API to make it more compatible and
performant to use in the upcoming MixerStage implementation. On the
other hand, this does not impact the main functionality to interact with
and schedule gain changes.

A follow up change will be submitted to introduce GainControls, similar
to ClockSnapshots, to be added into MixJobContext.

Bug: 87651

Multiply: audio-libprocessing-unittests
Change-Id: I93e4a128adbfc61805fc135ed45bda499052f97b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701890
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Alper Gungormusler <alperg@google.com>