[acpi] Add sleep state check to SetWakeDevice.
Ensure that a device is only set as a wake source if it is able to wake
the system from the requested sleep state.

Add tests for this and also add tests that power resources are turned on
when SetWakeDevice is called.

Change-Id: I87eb6f2ac323e8232089ba42cadd0c037b2a2f84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701397
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>