[identity] Add AccountPreAuthState to AccountHandlerControl.
We want to avoid giving account handler direct access to the
device-encrypted disk (and therefore the ability to exfiltrate large
volumes of account data if it were to be compromised). Previously we
solved this problem by giving account handler access to stash but stash
is deprecated and therefore we need a new solution.

This CL adds a new opaque data structure to the `AccountHandlerControl`
protocol, letting the account manager store state on the
device-encrypted disk on behalf of account handler.

Further work will be required both to persist this data structure in
AccountManager (fxbug.dev/104514) and also to use this data structure in
place of stash in AccountHandler (fxbug.dev/104516).

Bug: 104515
Change-Id: I2decf66892fbbb0299456d8c946724d72d67f979
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700919
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Angad Pal Singh <apsbhatia@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>