[starnix] Allow to create the init task without a filesystem
Most filesystem operation will require a Task to allow for permission
checking.

Before this CL, a filesystem was needed to create a task, which creates
a bootstrap problem.

This CL allows to create a task without a filesystem, and setting the
fielsystem later. This should only be used when creating the init task.

Change-Id: I1e2237691caa8be513e84261201bf48239a93046
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698382
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>