[scenic] Add flag to disable display composition
When true, the display compositor will not attempt to composite using
the display controller and will always fallback to the GPU compositor.
This is useful for testing.

This change also removes display_ids local variable, which was
initialized but never used.

TESTED=With flag both enabled and disabled, ensured everything rendered
correctly on workstation. Added logging to ensure each path was taken
when expected.

Change-Id: I8269918c670cd41b4eb9896966105c97bfdb543c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695966
Reviewed-by: Chris Fontas <cfontas@google.com>
Commit-Queue: David Worsham <dworsham@google.com>