[kernel][vm] Fix commit and resize race
If a commit is blocked on a page request and a resize invalidates the
range being waited on, the commit loop should not retry with a now
invalid range when unblocked. This used to work correctly until
fxr/693384 refactored the commit path and failed to bubble up the
termination condition from a lambda to the outer control loop.

The added test fails a range check debug assert without this fix.

Test: core-pager-writeback
Bug: 105012
Bug: 63989
Change-Id: Ic9f22ccdd8de789a746fb1c0493885b4049618b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703346
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>