[assembled_system] Parametrize kernel_zbi
Add the ability to select/override the kernel_zbi. This allows
assembled_system to interact with the zbi images generated by
zbi_test. This is meaningful for things like core-tests(kernel_zbi_test)
and early boot programs that want to propagate coverage data.

The default is '//zircon/kernel' which leaves the resulting image
unchanged if the parameter is not provided.

Bug: 101831
Test: fx build
Change-Id: Ib40506585e4cecbbce1c5c819de37c6624e852b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702802
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>