[fidl][cpp] Improve wire/natural conversion
Use the tricks from [1] to print an error when the user passes an
unknown type to the converter functions, as opposed to defaulting to a
no-op conversion.

Unfortunately that means we can't rely on the no-op template base case
and have to generate slightly more code to specialize the conversion
templates.

Add support for converting protocol endpoints and test them.

[1]: https://artificial-mind.net/blog/2020/10/03/always-false

Change-Id: Ide191523f8151be87e3f194d33f31e97ec6b564f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702527
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>