[security] Update bootfs scrutiny to rely on bootfs package index.
Implements a new scrutiny command for extracting the package index from
the bootfs. This is the first of three changes. The next change updates
the golden file scripts in vendor/google to use the extraction
functionality to generate golden files. The final change updates the
bootfs scrutinizer to use the golden files to validate the index.

Bug: 97517

Change-Id: I40e9be0ce7750fff51610ae58e8c3cc3d4ec13b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695920
Reviewed-by: Allison Pearce <ampearce@google.com>
Commit-Queue: Luke Nicholson <lukenicholson@google.com>