[kernel][vm] Handle gaps in DirtyPages
We can encounter gaps in DirtyPages if we raced with eviction and a
clean page was evicted before the DIRTY request could be resolved and
the page marked dirty. Fail the DirtyPages call in this case and let
waiters of the request retry looking up the page.

Test: core-pager-writeback
Bug: 63989
Multiply: core-pager-writeback
Run-All-Tests: true
Change-Id: Iba58ff4c585a9d6a6dcd9b7f9fe9794d40ffe257
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702322
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>