[kernel][coresight] Implicitly relativize access to component registers
Previously during a ROM table walk, we would access component registers
explicitly relative to the ROM table base. This prevented the (clean)
introduction of component class abstractions to facilitate that access,
as it would be hard to portably spell that offset in a way that would
makes sense when the component object is outside of the context of the
base ROM table; this also required a parallel `GetAt(uint32_t offset)`
API for all of the common component registers. In this change we reduce
that complexity by relativizing I/O to the current offset for each walk
node visited.

Test: k coresight walk on an Astro
Change-Id: I64f5d654a357ce999e1d148e20011702d7e9744a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437597
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>