[console-launcher] Retain dlog's virtcon handle
The current setup between dlog and console-launcher is inherently
error-swallowing. That's because dlog takes a handle to the virtcon
terminal as its stdout, and when that handle closes, virtcon gives you a
black screen. Retaining that handle allows us to continue to see dlog's
output after it dies.

Add logic to restart dlog if it dies, and do the same for the virtcon
shell, of which there is now only one. This same logic already exists
for the serial shell, and is reused here.

Change-Id: I9c6d8c2a619972b918ffa739b946791bc3c1015e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702964
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Tamir Duberstein <tamird@google.com>