[fidl][cpp] client side flexible method handling
Implement client side handling for flexible methods. This includes
sending the appropriate headers to indicate to the server that the
method is flexible and handling unpacking the result union used by
flexible methods.

This increases the complexity of the logic needed to correctly unpack
the result union. That logic was previously duplicated across
WireResult, WireUnownedResult (multiple times!), and
driver::WireUnownedResult. This was very error-prone as the
logic/templates needed to correctly unpack the result union (if any)
could be updated in one template without updating another template, and
there was a risk of having bugs in some code paths but not others, which
was very difficult to test for, especially since the architecture made
many of those paths very hard to exercise in unit test situations.

To deal with this, I created a BaseWireResult type which provides the
storage used for the wire result, the Unwrap and related accessors (if
applicable), and the logic for extracting the return
value/error/transport error from the result union. This ensures that
there is only one copy of that logic so it is the same for all
WireResult types.

Bug: 88366

Change-Id: Ic20ec7b0e262550dfe43a70b3a74be1462485994
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692530
Commit-Queue: Zachary Stewart <zstewart@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>