[virtualization] Implement SET_SCANOUT, FLUSH
Implement a scanout on top of a flatland view that will be presented
using GraphicalPresenter. For this implementation we still use a
hard-coded size for the virtual display and we don't yet attempt to
resize this to match the size of the view.

We disable the GetDisplayInfo device test since this test is now coupled
to having a functional flatland service. This test will be re-enabled
when we can properly create a view within the test.

Bug: 102870
Test: fx test virtio_gpu_unittests
Change-Id: I004ff0884b41ba775c853cd1d7af55707c59b672
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698090
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: David Hastings <dahastin@google.com>