[libc] Add trampoline for __hwasan_add_frame_record
Use trampolines similar to what's used for sancov functions for
circumventing PLT calls to __hwasan_add_frame_record. This is needed
in dynlink.c for instances where PLT calls cannot be made. The
strategy here is to define a local stub that __hwasan_add_frame_record
keeps a weak reference to so the linker elides the PLT entry and
resolves calls to the local stub.

This is also needed in allocate.c for a slightly different reason.
By the time we reach `map_block` for the first time via
`__init_main_thread`, `__hwasan_add_frame_record` will resolve to
the definition in the hwasan runtime, but since shadow call stack
is used in the runtime, we would page fault by accessing x18 which
is not setup yet intil after the call to `__init_main_thread`. In
allocate.c the stub is left empty to avoid a call to the runtime.

This CL includes some refactoring for reusing the sancov stub macro
for hwasan.

Bug: 57112
Change-Id: I0f1be46e4b6dbeec14f2a0b3e474b2528ddf352f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702145
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>