[ffx][package] Fix ffx package build --depfile
According to the [gn reference], all files that are read by the action
must be included in the depfile. This changes `ffx package build
--depfile` to make sure we include the files we'll be putting into the
far package.

In addition, our gn will expresses that `ffx package build` produces
`package_manifest.json`, so we don't need to include that in the depfile.

In addition, this makes sure that the entries in the depfile are sorted
and deduplicated to make sure we generate consistent files.

[gn reference]: https://gn.googlesource.com/gn/+/master/docs/reference.md#var_depfile

Change-Id: I3a6eeb7237e22ab5a8cb3a1e70e874d4736d8a25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703722
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>