[ffx] Launch an editor from ffx emu
One of the biggest issues with the initial version of `ffx emu` is the
difficulty users found with modifying configuration. We put forward
go/ffx-emu-config-design as an improvement for this situation.

This CL implements the most crucial aspect of this design, which is
actually launching the editor on a launch configuration object. We do
this via an "--edit" flag, which starts up the user's default editor
(via the $EDITOR environment flag) or '/usr/bin/nano' by default. The
resulting value overwrites the existing FlagData object for the emulator
engine, as long as the editor process returns a successful error code.

Bug: 98268

Change-Id: I4adcab34e8e597fbdad1d61ff3c6058ddf9c6f10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701233
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>