[ffx][debugger] Add signal handlers to --agent-only
Add signal handlers for SIGTERM and SIGINT to the --agent-only mode of
operation for the debug plugin. This allows all current uses of the
plugin to continue working while also enabling other processes to launch
DebugAgent in a typical way and also properly clean up the sockets on
the host side.

Change-Id: I4dbfc10a301ed5e5e9a3bc88ea21077f60431e6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702540
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>