[kernel] Add restricted aspace to process dispatcher
This is the address space that will be used by threads that are running
in restricted mode.

Also moves the normal aspace to the shareable process state, since this
will be shareable in the future (the restricted aspace will not be).

This change does not introduce any way for clients to create processes
with restricted address spaces, or use it in any meaningful way.

A follow up change will add the ability to create processes that both
share their "shareable process state" with other processes, and contain
a restricted address space.

Bug: 91399, 98922

Change-Id: I37f30d1d4ff952f79dc161cb7171e13d4933237d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673428
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>